requirements.txt
playwright==1.41.0
python-dotenv==1.0.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0
pychromecast==13.1.0
gtts==2.5.0

